I have completed javascript
JavaScript is a versatile programming language commonly used in web development to enhance the user experiences on the internet. It is primarily utilised for implementing functionalities within our web applications. JavaScript is often hailed as a the "language of the web", is the most important part of modern web development.